+++ CHANGELOG.md
@@ -8,6 +8,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## `0.1.10` - October 4th, 2023
+
+### Fixed
+
+-   Fixed prerelease versions of rojo being detected as incompatible
+
 ## `0.1.9` - August 25th, 2023
 
 ### Added